feat(rand-8217): updated heuristic-result

+++ src/ShutUpHusky/Heuristics/HeuristicResult.cs
@@ -1,7 +1,7 @@
 namespace ShutUpHusky.Heuristics;
 
 internal record HeuristicResult {
-    public required int Priority { get; init; }
+    public required double Priority { get; init; }
     public required string Value { get; init; } = string.Empty;
     public string? After { get; init; } = null;
 }
